Use S3MultipartUploadStream in S3NFileSystem.
Use new S3MultipartUploadStream as the underlying OutputStream for our
version of the S3N FileSystem. S3MultipartUploadStream uses the Amazon
multi-part protocol, and ,thus, does not need to pre-buffer the file to
the local FS (like the original S3NFileSystem Impl) before initiating a
transfer up to S3. Also, S3MultipartUploadStream is more resilient to
failures (thanks to its use of the multi-part protocol).